// BN/BN-frontend/src/models/RN_Menues.ts
export interface IMenues {
  _id: string,
  restaurantName: string;
  restaurantAddress: string;
  restaurantPostalCode: string;
  restaurantCity: string;
  restaurantPhone: string;
  restaurantURL: string;
  restaurantContact: string;
  menuDescription: string;
  restaurantUserName: string;
  restaurantPassword: string;
  menuPrize: number;
  
}


export interface INewMenues {
  restaurantName: string;
  restaurantAddress: string;
  restaurantPostalCode: string;
  restaurantCity: string;
  restaurantPhone: string;
  restaurantURL: string;
  restaurantContact: string;
  menuDescription: string;
  restaurantUserName: string;
  restaurantPassword: string;
  menuPrize: number;
 
}